@@ -5,17 +5,24 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "run-p type-check \"build-only {@}\" --",
+    "build": "vite build",
     "preview": "vite preview",
-    "test:unit": "vitest",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --build --force",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
-    "format": "prettier --write src/"
+    "test:unit": "vitest run",
+    "test:unit:watch": "vitest",
+    "lint": "run-s lint:*",
+    "lint:code-quality": "eslint",
+    "lint:format": "prettier --check .",
+    "lint:types": "vue-tsc --build --noEmit",
+    "fix": "run-s fix:*",
+    "fix:code-quality": "eslint --fix",
+    "fix:format": "prettier --write ."
+  },
+  "engines": {
+    "node": "23.11.1"
   },
   "dependencies": {
     "@auth0/auth0-vue": "^2.3.3",
-    "@dpa-id-components/dpa-shared-components": "^9.4.3",
+    "@dpa-id-components/dpa-shared-components": "16.0.0",
     "@dpa-it/dpa-id-partner-components-vue": "^3.0.8",
     "@vitejs/plugin-basic-ssl": "^1.1.0",
     "axios": "^1.7.2",
@@ -24,25 +31,21 @@
     "vue-router": "^4.3.0"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.8.0",
-    "@tsconfig/node20": "^20.1.4",
+    "@tsconfig/node22": "22.0.2",
     "@types/jsdom": "^21.1.6",
-    "@types/node": "^20.12.5",
+    "@types/node": "22.16.3",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitejs/plugin-vue-jsx": "^3.1.0",
-    "@vue/eslint-config-prettier": "^9.0.0",
-    "@vue/eslint-config-typescript": "^13.0.0",
     "@vue/test-utils": "^2.4.5",
     "@vue/tsconfig": "^0.5.1",
     "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-plugin-vue": "^9.23.0",
+    "eslint": "9.30.1",
     "jsdom": "^24.0.0",
-    "npm-run-all2": "^6.1.2",
+    "npm-run-all2": "^6.2.6",
     "postcss": "^8.4.38",
     "prettier": "^3.2.5",
     "tailwindcss": "^3.4.4",
-    "typescript": "~5.4.0",
+    "typescript": "5.8.3",
     "vite": "^5.2.8",
     "vue-tsc": "^2.0.11"
   }
