@@ -3,28 +3,27 @@ import "source-map-support/register";
 import * as cdk from "aws-cdk-lib";
 import { Configuration } from "../src/config/configuration";
 import { ConfigManager } from "../src/config/config-manager";
-import {DpaIdAuth0SpaDemoStack} from "../src/stacks/dpaidauth0spademo/dpa-id-auth0-spa-demo-stack";
+import { DpaIdAuth0SpaDemoStack } from "../src/stacks/dpaidauth0spademo/dpa-id-auth0-spa-demo-stack";
 
 const app = new cdk.App();
 
 const stage = process.env.STAGE || null;
-console.log("STAGE : "+stage)
+console.log("STAGE : " + stage);
 if (!stage) throw new Error("STAGE is not set");
 
 const config: Configuration = ConfigManager.createConfig(stage);
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const dpaIdAuth0SpaDemoStack = new DpaIdAuth0SpaDemoStack(
-    app,
-    "dpa-id-auth0-spa-demo-" + config.settings.deploymentSettings.stageSuffix,
-    {
-        env: config.env,
-        settings: config.settings,
-        tags: {
-            project: "dpa-id-auth0-spa-demo",
-            app: "dpa-id-auth0-spa-demo",
-        },
-        description:
-            "dpaId Auth0 SPA Demo Service Stack."
-    }
+  app,
+  "dpa-id-auth0-spa-demo-" + config.settings.deploymentSettings.stageSuffix,
+  {
+    env: config.env,
+    settings: config.settings,
+    tags: {
+      project: "dpa-id-auth0-spa-demo",
+      app: "dpa-id-auth0-spa-demo",
+    },
+    description: "dpaId Auth0 SPA Demo Service Stack.",
+  },
 );
