@@ -1,15 +1,21 @@
-import { Construct } from 'constructs'
-import {Effect, IRole, PolicyStatement, Role, ServicePrincipal} from 'aws-cdk-lib/aws-iam'
-import {Stack} from "aws-cdk-lib";
+import { Construct } from "constructs";
+import {
+  Effect,
+  IRole,
+  PolicyStatement,
+  Role,
+  ServicePrincipal,
+} from "aws-cdk-lib/aws-iam";
+import { Stack } from "aws-cdk-lib";
 
 export class DpaIdAuth0SpaDemoRole {
-  public readonly instance: IRole
+  public readonly instance: IRole;
 
   constructor(scope: Construct) {
-    const role = new Role(scope, 'DpaIdAuth0SpaDemo-Role', {
-      assumedBy: new ServicePrincipal('ecs-tasks.amazonaws.com')
+    const role = new Role(scope, "DpaIdAuth0SpaDemo-Role", {
+      assumedBy: new ServicePrincipal("ecs-tasks.amazonaws.com"),
     });
-    this.instance = role
+    this.instance = role;
 
     DpaIdAuth0SpaDemoRole.allowAccessToParameterStore(role);
     this.instance = role;
@@ -21,14 +27,13 @@ export class DpaIdAuth0SpaDemoRole {
     const account = parentStack.account;
 
     role.addToPolicy(
-        new PolicyStatement({
-          resources: [
-             `arn:aws:ssm:${region}:${account}:parameter/config/dpa-id-auth0-spa-demo*`,
-          ],
-          actions: ["ssm:GetParametersByPath"],
-          effect: Effect.ALLOW,
-        })
+      new PolicyStatement({
+        resources: [
+          `arn:aws:ssm:${region}:${account}:parameter/config/dpa-id-auth0-spa-demo*`,
+        ],
+        actions: ["ssm:GetParametersByPath"],
+        effect: Effect.ALLOW,
+      }),
     );
   }
-
 }
