@@ -5,10 +5,10 @@
 </template>
 
 <script setup lang="ts">
-import { RouterView } from 'vue-router'
-import { onMounted } from 'vue'
-import { useRouter } from 'vue-router'
+import { RouterView } from "vue-router";
+import { onMounted } from "vue";
+import { useRouter } from "vue-router";
 
-const router = useRouter()
-onMounted(() => router.push('/'))
+const router = useRouter();
+onMounted(() => router.push("/"));
 </script>
