@@ -6,6 +6,6 @@
 </template>
 
 <script setup lang="ts">
-import DemoProfile from '@/components/DemoProfile.vue'
-import DemoHeader from '@/components/DemoHeader.vue'
+import DemoProfile from "@/components/DemoProfile.vue";
+import DemoHeader from "@/components/DemoHeader.vue";
 </script>
